[REFACTOR] Rename posts reducer resources to ensure clarity and differentiation
