+++ src/modules/posts/create/components/steps/default/footer/CreatePostDefaultStepFooter.tsx
@@ -4,7 +4,7 @@ import { useTranslations } from 'next-intl';
 import { IS_PRODUCTION } from '@/constants/environment';
 import { useAppDispatch, useAppSelector } from '@/lib/store/hooks';
 import { CardItem } from '@/modules/posts/create/assets/types';
-import { setCreatePostStep } from '@/modules/posts/create/reducers/steps/createPostPostReducer';
+import { setCreatePostStep } from '@/modules/posts/create/reducers/createPostPostReducer';
 
 import { CreatePostDefaultStepFooterItem } from './item/CreatePostDefaultStepFooterItem';
