+++ src/modules/posts/create/reducers/createPostReducer.ts
@@ -1,17 +1,17 @@
 import { combineReducers } from '@reduxjs/toolkit';
 
+import createPostPostReducer from './createPostPostReducer';
 import createPostCheckInStepReducer from './steps/createPostCheckInStepReducer';
 import createPostFeelingsStepReducer from './steps/createPostFeelingsStepReducer';
 import createPostGifsStepReducer from './steps/createPostGifsStepReducer';
 import createPostMediaStepReducer from './steps/createPostMediaStepReducer';
-import createPostPostReducer from './steps/createPostPostReducer';
 
 const createPostReducer = combineReducers({
+  createPostPost: createPostPostReducer,
   createPostCheckInStep: createPostCheckInStepReducer,
   createPostFeelingsStep: createPostFeelingsStepReducer,
   createPostGifsStep: createPostGifsStepReducer,
   createPostMediaStep: createPostMediaStepReducer,
-  createPostPost: createPostPostReducer,
 });
 
 export default createPostReducer;
